supervisor: Register on-finalized ACKs to fork_point()-s
This is more consistent with the wait() API which waits for a pid to terminate,
not for a pid's current executed binary to finish running.
