+++ src/firebuild/process.cc
@@ -921,24 +921,11 @@ bool Process::any_child_not_finalized() {
   return false;
 }
 
-void Process::maybe_send_on_finalized_ack() {
-  if (on_finalized_ack_id_ != -1) {
-    assert(on_finalized_ack_fd_ != -1);
-    ack_msg(on_finalized_ack_fd_, on_finalized_ack_id_);
-    on_finalized_ack_id_ = -1;
-    on_finalized_ack_fd_ = -1;
-  }
-}
-
 /**
  * Finalize the current process.
  */
 void Process::do_finalize() {
   TRACKX(FB_DEBUG_PROC, 1, 1, Process, this, "");
-
-  /* Now we can ack the previous system()'s second message,
-   * or a pending pclose() or wait*(). */
-  maybe_send_on_finalized_ack();
 
   reset_file_fd_pipe_refs();
